Move nexus and generator files to top dir.
